Split up functions with the same file structure as physis
It was a pain working in a ~1.3K LoC file, but there's no point in
putting it all in one file. Now libphysis copies the same file structure
as upstream physis, so it's extremely easy to switch between them when
working.

This does not affect the header generation, everything is still in one
file.